% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{assign_color}
\alias{assign_color}
\title{Core color mapping function for Spaco2.}
\usage{
assign_color(
  cluster_distance_matrix = data.frame(),
  colorblind_type = c("none", "protanopia", "deuteranopia", "tritanopia", "general"),
  palette = NULL,
  image_palette = NULL,
  ...
)
}
\arguments{
\item{cluster_distance_matrix}{A matrix representing the distances between clusters.A DataFrame with unique cluster names as \code{index} and \code{columns}, which contains a distance adjacent matrix for clusters, representing the dissimilarity between clusters.}

\item{colorblind_type}{A character vector specifying the type of colorblindness
to accommodate. Options include "none", "protanopia", "deuteranopia", "tritanopia",
and "general". Default is "none".}

\item{palette}{An optional vector of color values (in hex format). If provided,
this palette will be used and \code{image_palette} will be ignored.Defaults to None.}

\item{image_palette}{An optional image (in a format compatible with R) used to
extract a color palette. Ignored if \code{palette} is provided.Defaults to None.}

\item{mapping_args}{A list of additional arguments to pass to the \code{map_graph} function.}

\item{embed_args}{A list of additional arguments to pass to the \code{embed_graph} function.}
}
\value{
A named vector where names are cluster identifiers and values are the assigned
hex color codes.
}
\description{
Assign Colors to Clusters Based on Distance Matrix.
This function assigns colors to clusters based on a given distance matrix.
It supports colorblind-friendly options and can automatically generate a color palette,
use a predefined palette, or extract colors from an image.
Spaco2 provides 3 basic color mapping mode in this function:
\enumerate{
\item Optimize the mapping of a pre-defined color palette.
\item Extract colors from image.
\item Automatically generate colors within colorspace.
}
}
