% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{spatial_distance}
\alias{spatial_distance}
\alias{perceptual_distance}
\title{calculate spatial interlacement distance graph for cell clusters}
\usage{
spatial_distance(
  cell_coordinates,
  cell_labels,
  neighbor_weight = 0.5,
  radius = 90,
  n_neighbors = 16,
  n_cells = 3
)

perceptual_distance(
  colors,
  colorblind_type = c("none", "protanopia", "deuteranopia", "tritanopia", "general")
)
}
\arguments{
\item{cell_coordinates}{a list like object containing spatial coordinates for each cell}

\item{cell_labels}{a list like object containing cluster labels for each cell}

\item{neighbor_weight}{cell weight to calculate cell neighborhood. Defaults to 0.5}

\item{radius}{radius used to calculate cell neighborhood. Defaults to 90}

\item{n_neighbors}{k for KNN neighbor detection. Defaults to 16}

\item{n_cells}{nly calculate neighborhood with more than \code{n_cells}. Defaults to 3}

\item{colors}{a list of colors (in hex)}

\item{colorblind_type}{optional parameter}
}
\value{
a Data.Frame with unique cluster names as \code{index} and \code{columns},
which contains interlacement distance between clusters

a data.frame with unique colors (in hex) as index and columns,which contains perceptual distance between colors
}
\description{
Function to calculate spatial interlacement distance graph for cell clusters, where
we define the interlacement distance as the number of neighboring cells between two cluster

See 'color_difference_rgb' for details
}
