% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matrix_distance}
\alias{matrix_distance}
\alias{hex_to_rgb}
\alias{rgb_to_hex}
\alias{lab_to_hex}
\alias{rgb_to_lms_img}
\alias{lms_to_rgb_img}
\alias{color_difference_rgb}
\alias{get_bin_color}
\alias{palette_min_distance}
\alias{color_score}
\alias{simulate_cvd}
\alias{extract_palette}
\title{Calculate the distance between two matrices}
\usage{
matrix_distance(matrix_x, matrix_y, metric = "manhattan")

hex_to_rgb(hex_code)

rgb_to_hex(rgb_code)

lab_to_hex(lab_code)

rgb_to_lms_img(img)

lms_to_rgb_img(img)

color_difference_rgb(color_x, color_y)

get_bin_color(bin_number)

palette_min_distance(palette)

color_score(lab_color, color_count, palette, wn)

simulate_cvd(palette_hex, colorblind_type)

extract_palette(
  reference_image,
  n_colors,
  colorblind_type,
  l_range = c(20, 85),
  trim_percentile = 0.03,
  max_iteration = 20,
  verbose = FALSE
)
}
\arguments{
\item{matrix_x}{matrix x}

\item{matrix_y}{matrix y}

\item{metric}{metric used to calculate distance. Defaults to manhattan}

\item{rgb_code}{RGB channel value}

\item{lab_code}{CIE Lab color value}

\item{img}{lms image}

\item{color_x}{color_x}

\item{color_y}{color_x}

\item{bin_number}{numbered bin color}

\item{palette}{palette}

\item{lab_color}{lab_color}

\item{color_count}{color_count}

\item{wn}{float}

\item{palette_hex}{palette_hex}

\item{colorblind_type}{colorblind type}

\item{reference_image}{reference_image}

\item{n_colors}{colors}

\item{l_range}{intergrate}

\item{trim_percentile}{trim_percentile}

\item{max_iteration}{max_iteration}

\item{verbose}{verbose}

\item{hex_codehex}{string representing a RGB color}
}
\value{
the distance between two matrices

integer values for RGB channels

color hex string

color hex string

lms image matrix

RGB image matrix

the perceptual difference between two colors

Lab values for the centroid color of this bin

the min distance

color_score

rgb_to_hex

lab_to_hex
}
\description{
Calculate the distance between two matrices

Convert hex string to RGB value

Convert RGB value (0~255) to hex string

Convert CIE Lab color value to hex string

Convert RGB image matrix (0~255) to lms image matrix

Convert lms image matrix to RGB image matrix

Calculate the perceptual difference between colors

Revert bin number in \code{extract_palette} function to Lab values

Calculate the minimal distance within a Lab palette

Score color replacement

Extract palette from image
}
